requirements.txt
python-decouple==3.8
aiogram==3.17.0
aiosqlite==0.21.0